+++ src/constants/cache.constant.ts
@@ -8,7 +8,7 @@
 export const CACHE_PREFIX = '__nodepress_cache_'
 export const OPTION = CACHE_PREFIX + 'option'
 export const ALL_TAGS = CACHE_PREFIX + 'all-tags'
-export const HOT_ARTICLES = CACHE_PREFIX + 'hot-articles'
+export const HOTTEST_ARTICLES = CACHE_PREFIX + 'hottest-articles'
 export const ARCHIVE = CACHE_PREFIX + 'archive'
 export const TODAY_VIEWS = CACHE_PREFIX + 'today-views'